+++ pkg/storage/stores/shipper/indexshipper/tsdb/head_manager_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/grafana/loki/v3/pkg/storage/config"
 	"github.com/grafana/loki/v3/pkg/storage/stores/shipper/indexshipper"
 	"github.com/grafana/loki/v3/pkg/storage/stores/shipper/indexshipper/tsdb/index"
+	"github.com/grafana/loki/v3/pkg/storage/types"
 	"github.com/grafana/loki/v3/pkg/validation"
 )
 
@@ -443,8 +444,8 @@ func TestBuildLegacyWALs(t *testing.T) {
 			Configs: []config.PeriodConfig{
 				{
 					Schema:     "v11",
-					IndexType:  config.TSDBType,
-					ObjectType: config.StorageTypeFileSystem,
+					IndexType:  types.TSDBType,
+					ObjectType: types.StorageTypeFileSystem,
 					IndexTables: config.IndexPeriodicTableConfig{
 						PeriodicTableConfig: config.PeriodicTableConfig{
 							Prefix: "index_",
@@ -454,8 +455,8 @@ func TestBuildLegacyWALs(t *testing.T) {
 				{
 					Schema:     "v11",
 					From:       config.DayTime{Time: timeToModelTime(secondStoreDate)},
-					IndexType:  config.TSDBType,
-					ObjectType: config.StorageTypeFileSystem,
+					IndexType:  types.TSDBType,
+					ObjectType: types.StorageTypeFileSystem,
 					IndexTables: config.IndexPeriodicTableConfig{
 						PeriodicTableConfig: config.PeriodicTableConfig{
 							Prefix: "index_",
@@ -467,8 +468,8 @@ func TestBuildLegacyWALs(t *testing.T) {
 			Configs: []config.PeriodConfig{
 				{
 					Schema:     "v12",
-					IndexType:  config.TSDBType,
-					ObjectType: config.StorageTypeFileSystem,
+					IndexType:  types.TSDBType,
+					ObjectType: types.StorageTypeFileSystem,
 					IndexTables: config.IndexPeriodicTableConfig{
 						PeriodicTableConfig: config.PeriodicTableConfig{
 							Prefix: "index_",
@@ -478,8 +479,8 @@ func TestBuildLegacyWALs(t *testing.T) {
 				{
 					Schema:     "v12",
 					From:       config.DayTime{Time: timeToModelTime(secondStoreDate)},
-					IndexType:  config.TSDBType,
-					ObjectType: config.StorageTypeFileSystem,
+					IndexType:  types.TSDBType,
+					ObjectType: types.StorageTypeFileSystem,
 					IndexTables: config.IndexPeriodicTableConfig{
 						PeriodicTableConfig: config.PeriodicTableConfig{
 							Prefix: "index_",
@@ -491,8 +492,8 @@ func TestBuildLegacyWALs(t *testing.T) {
 			Configs: []config.PeriodConfig{
 				{
 					Schema:     "v13",
-					IndexType:  config.TSDBType,
-					ObjectType: config.StorageTypeFileSystem,
+					IndexType:  types.TSDBType,
+					ObjectType: types.StorageTypeFileSystem,
 					IndexTables: config.IndexPeriodicTableConfig{
 						PeriodicTableConfig: config.PeriodicTableConfig{
 							Prefix: "index_",
@@ -502,8 +503,8 @@ func TestBuildLegacyWALs(t *testing.T) {
 				{
 					Schema:     "v13",
 					From:       config.DayTime{Time: timeToModelTime(secondStoreDate)},
-					IndexType:  config.TSDBType,
-					ObjectType: config.StorageTypeFileSystem,
+					IndexType:  types.TSDBType,
+					ObjectType: types.StorageTypeFileSystem,
 					IndexTables: config.IndexPeriodicTableConfig{
 						PeriodicTableConfig: config.PeriodicTableConfig{
 							Prefix: "index_",